CLN: ChatPage 화면 추가
