+++ app/ChatPage.tsx
@@ -1,55 +1,35 @@
 'use client';
 
 import { useState } from 'react';
-import { ChatLogService } from '@/business/ChatLogService';
-import { OpenAiRole, IOpenAiParam } from '@/types/OpenAiParam';
 import Header from '@/components/Header';
-import Chat from '@/components/Chat';
 import ChatMessageInput from '@/components/ChatMessageInput';
 
 export default function ChatPage() {
   const [isTyping, setIsTyping] = useState(false);
-  const [chatLogList, setChatLogList] = useState<IOpenAiParam[]>([]);
 
   const onSubmit = async (chatLog: string) => {
     console.log('==========handleSubmit==========');
-    setIsTyping(true);
-
-    const userChat: IOpenAiParam[] = [...chatLogList, { role: OpenAiRole.user, content: chatLog }];
-    setChatLogList(userChat);
-
-    const chatLogService = new ChatLogService();
-    const response = await chatLogService.getAnswer(userChat);
-    setIsTyping(false);
-
-    const assistantChat: IOpenAiParam[] = [
-      ...userChat,
-      {
-        role: OpenAiRole.assistant,
-        content: response,
-      },
-    ];
-    setChatLogList(assistantChat);
+    console.log(chatLog);
   };
 
   return (
     <div className="grid h-screen w-full">
       <div className="flex flex-col">
         <Header />
-        <main className="relative flex h-full min-h-[50vh] flex-col bg-white p-4 lg:col-span-2">
-          <div className="flex-1 w-full max-w-2xl m-auto">
-            <div className="space-y-4">
-              {chatLogList.map((chat, index) => (
-                <Chat key={chat.content + index} role={chat.role} content={chat.content}></Chat>
-              ))}
+        <main className="relative flex h-full min-h-[50vh] flex-col bg-white lg:col-span-2">
+          <div className="w-full max-w-2xl m-auto">
+            <div className="flex flex-col items-center w-full h-full space-y-4">
+              <h2 className="text-3xl font-bold text-center text-gray-800 tracking-wide leading-tight">
+                궁금한 것은 무엇이든 물어보세요.
+              </h2>
+              <div className="w-full bg-white border p-5 rounded-2xl">
+                <div className="max-w-2xl m-auto">
+                  <ChatMessageInput onSubmit={onSubmit} isTyping={isTyping} />
+                </div>
+              </div>
             </div>
           </div>
         </main>
-        <div className="sticky bottom-0 w-full py-2 border-t bg-white">
-          <div className="max-w-2xl m-auto">
-            <ChatMessageInput onSubmit={onSubmit} isTyping={isTyping} />
-          </div>
-        </div>
       </div>
     </div>
   );
